Fixing a regression bug
When remove the ~ (tilde) syntax from package.json, it was not replaced with thew new ^ syntax.
Version 0.2.1 therefore was using 0.1.3 of each-async, while Version 1.0 regressed to 0.1.0 of each-async.

- Adds ^ syntax to each-async
- Sets minimum each-async to 0.1.3

Further; UTF8 prior to node@0.10.29 would incorrectly handle utf8 conversion from non-compatible sources (like a binary image).
To keep this project from corrupting previously hashes, sniffing of the node version and flags must occur to give the developer
the exact hash they are expecting.

It should not be encouraged to use UFT8 for binary files; but for backward compatibility this PR implements all approaches with passing tests.